Change ImagePixelBox to use size_t.

+++ SerialPrograms/Source/PokemonSwSh/Inference/PokemonSwSh_TypeSymbolFinder.cpp
@@ -33,21 +33,21 @@ size_t distance_sqr(const ImagePixelBox& a, const ImagePixelBox& b){
         return 0;
     }
 
-    ptrdiff_t dist_x = 0;
+    size_t dist_x = 0;
     if (!overlap_x){
         dist_x = a.max_x < b.min_x
             ? b.min_x - a.max_x
             : a.min_x - b.max_x;
     }
 
-    ptrdiff_t dist_y = 0;
+    size_t dist_y = 0;
     if (!overlap_y){
         dist_y = a.max_y < b.min_y
             ? b.min_y - a.max_y
             : a.min_y - b.max_y;
     }
 
-    return (size_t)dist_x*dist_x + (size_t)dist_y*dist_y;
+    return dist_x*dist_x + dist_y*dist_y;
 }
